requirements_chirho.txt
# For God so loved the world, that He gave His only begotten Son, that all who believe in Him should not perish but have everlasting life
aiosqlite
discord.py
icecream
python-dotenv

